Add generate-fontimage.sh
The new tool "generate-fontimage.sh" is to produce a thumbnail image.
This tool replaces the main logic of font-image.yml and it is also used
by build-dist.sh internally.

build-dist.sh is now to produce fontimage.png everytime and to store it
in the npm package. Then, README.md in the npm package (i.e. a copy of
README-npm.md) is now to refer to the produced fontimage.png, instead of
referring to /fontimage.png at "main" branch in GitHub. Using the main
branch URL was problematic because it meant README in any version of the
npm package always refers to the latest image, which can be different
from what the version actually provides.